+++ testproject/Assets/Tests/Manual/InSceneObjectParentingTests/InSceneParentChildHandler.cs
@@ -174,6 +174,7 @@ namespace TestProject.RuntimeTests
                 {
                     ServerRootParent = this;
                 }
+                ServerRelativeInstances.Add(NetworkObjectId, this);
             }
             else
             {